bot_tg/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
openai==1.101.0
orjson==3.11.2
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
rsa==4.9.1
sniffio==1.3.1
socksio==1.0.0
Telethon==1.40.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvloop==0.21.0
